Add Planner
